// lib/api/generated/model/promptActivityRow.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Pear API
 * OpenAPI spec version: 1.0.0
 */
import type { PromptActivityStatus } from './promptActivityStatus';

export interface PromptActivityRow {
  id: string;
  daterId: string;
  daterName: string;
  promptQuestion: string;
  message: string;
  status: PromptActivityStatus;
  createdAt: string;
}